Fix start and due dates values for work items
On https://gitlab.com/gitlab-org/gitlab/-/merge_requests/170497, we
noticed that we were creating `WorkItems::DatesSource` with empty
values. As a side effect, sometimes the dates were not showing properly
on some parts of the UI.

To fix that, instead of only falling back from `WorkItems::DatesSource`
to the `WorkItem` dates fields, we need to check if the
`WorkItems::DateSource` has any values.

Related to:
- https://gitlab.com/gitlab-org/gitlab/-/issues/500364
- https://gitlab.com/gitlab-org/gitlab/-/issues/500579

Changelog: fixed
EE: true