Remove instead of fade to-do items that get marked as (un)done
This removes the custom fade effect from todos that just got marked as
done, or undone.

Instead, a full Apollo `refetch` gets triggered, which, when finished,
removes the todo from the list.

To make this effect smoother, Vue list transitions are used.

It was also required to change the Apollo cache strategy. Otherwise,
on a long, paginated list, checking (removing) an item would not load
in the next item to fill up that page. Nor would that removed item show
up on the other tab if that tab was already cached earlier.