Fix errors from bundle install when cache is not available
When cache does not exist, it'll try to install rugged,
and building and installing rugged requires cmake, which
the default ruby image doesn't have.

This script actually doesn't use any gems, only built-in libs.
We don't need to install anything anyway, so we don't use bundler,
and we don't need to use cache