Avoid use of negative lookahead in JSON Schema regular expressions
The latest version of the JSON Schema specification mentions the fact
that there are many different regular expression variants and says:

> given the high disparity in regular expression constructs support,
> schema authors SHOULD limit themselves to the following regular expression tokens:

(See https://json-schema.org/draft/2020-12/draft-bhutton-json-schema-01#name-regular-expressions)

One example of such a variant is Go's regular expression evaluator
which, despite, and in fact because of, some limitations, such as no
backreferences or lookahead, has some technical advantages such as a
guaranteed non-exponential runtime (see
https://swtch.com/~rsc/regexp/regexp1.html for some background).

Our JSON Schema checker uses that implementation for matching regular
expressions, which means that it cannot check the regular expressions in
`ci.json` that contain negative lookahead assertions.

Given the above, this PR changes the schema to avoid using negative
lookahead assertions in its regular expressions.

The `cache_item.key` pattern is the easiest of the two. It's easy to
rewrite as a functionally identical pattern that doesn't use negative
lookahead. The lookahead says "there is no / character in the string",
so the rest of the pattern can be changed to allow only characters which
aren't `/`.

The `include_item` pattern is a little more complex, as it doesn't
appear to be easily translatable to an equivalent without negative
lookahead. My reading of the sense of the current pattern's intent is:

- the URL must refer to a YAML file
- it can be an HTTP or HTTPS URL but no other scheme.

We can write this using an overall pattern that requires the URL to be a
YAML file, along with a requirement that it's either an HTTP URL or a
URL without a scheme.

To my mind, this is actually a more direct understandable way to express
the constraint even if we weren't trying to eliminate negative
lookahead.

When translating, I also realised that there is a probable bug in the
existing pattern: it allows `xx.yaml` but not `x.yaml`. I doubt that's
intended behavior, I've altered it to allow single-letter basenames too.

I've done some checking to make sure the change really does match the
same strings, and I've made a couple of JSON schemas where that
behaviour can be explored: in each case, the schema checks that for an
array of input strings, both the old and the new schemas either both
pass or both fail. We can see that there's a discrepancy as expected
when `x.yaml` is included in the set of test strings.

Sanity checker for the `cache_item.key` change: https://www.jsonschemavalidator.net/s/7SzAC4ed
Sanity checker for the include_item` change: https://www.jsonschemavalidator.net/s/GEyHoxVd

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>