Enable internal Repository X-Ray service
On projects that have Duo code suggestions enabled,
the Repository X-Ray is automatically executed (via
background sidekiq job) when a new commit is pushed to
the default branch of the repo. The X-Ray scans the
repository for supported dependency manager configuration
files and extracts a list of libraries from them. The list
is saved to 'xray_reports' table where it is later used
to augment the context for the Code Generation prompt.

Feature flag removed 'ai_enable_internal_repository_xray_service'

Changelog: added