Stop blocking pre-existing policy rules as unenforceable
Currently we're triggering a policy bot comment for
unenforceable rules via `UnenforceablePolicyRulesNotificationService`
to display configuration errors to the user.
These are based on missing scans in the pipeline.

Pre-existing policy rules are evaluated separately and
don't require pipeline to be evaluated.
We shouldn't include them in this evaluation,
otherwise they conflict with the
`SyncPreexistingStatesApprovalRulesWorker` and cause approvals
to be reset, even if there is no policy
violation of the pre-existing rules.

Changelog: fixed
EE: true