Refactor Gitlab::CodeOwners::entries_for_merge_request
This change makes the method more readable and removes some unnecessary
checks.

We have a foreign key constraint on merge_requests.source_project_id so
we don't need to load the source_project to check it exists, we can
simply check if the id value is stored in the database.
https://gitlab.com/gitlab-org/gitlab/-/blob/8fd220fe55dd76f2301d65a20f775dcf9655db32/db/structure.sql#L38367-38368

We also have a not-null contraint on the source_branch column so we can
completely skip that check.
https://gitlab.com/gitlab-org/gitlab/-/blob/8fd220fe55dd76f2301d65a20f775dcf9655db32/db/structure.sql#L14337

merge_request.project is an alias of merge_request.target_project so the
feature_available check was duplicated. Now removed.
https://gitlab.com/gitlab-org/gitlab/-/blob/8fd220fe55dd76f2301d65a20f775dcf9655db32/app/models/merge_request.rb#L505
https://gitlab.com/gitlab-org/gitlab/-/blob/8fd220fe55dd76f2301d65a20f775dcf9655db32/ee/lib/gitlab/code_owners.rb#L40
https://gitlab.com/gitlab-org/gitlab/-/blob/8fd220fe55dd76f2301d65a20f775dcf9655db32/ee/lib/gitlab/code_owners.rb#L70