Allow interrupted RelationBatchExportWorker to bypass limit
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/169122, a limit
was established to restrict the number of concurrent
RelationBatchExportWorker jobs. However, when jobs are interrupted,
they are only retried after being classified as timed out, which can
take a considerable amount of time.

This change updates the limit to be bypassed when the batch is already
in a started state, allowing interrupted jobs to be retried immediately.

Changelog: changed