Fix LLM AI client not returning a HTTP response with 204 responses
Previously `Llm::AiGateway::Client#request` returned `nil` if the AI
Gateway return a `204 No Content` response. However, this makes it
impossible to discern whether the request was successful or whether
the server returned a 5xx error.

This was happening because `run_retry_with_exponential_backoff`
returned `nil` if the body were blank. To fix this, return the
HTTParty response even the body is blank, but ensure the callers
handle this.

Changelog: fixed