Fix performance of To-Do GraphQL API when sorting by created
Todos are created sequentially, so id is a reasonable proxy for
created_at. We use this fact to optimize the performance of the finder.
We probably could have adjusted the GraphQL enum to allow sorting by id
instead and used that in the Frontend, but this would take multiple
release cycles and not make it faster for the people who are using
CREATED_DESC as a filter in manual queries.

See https://gitlab.com/gitlab-org/gitlab/-/work_items/502920