Emails on Push: fix wrong number of changed files
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/17532

**Problem**

When there are to many changes in the commit, `#diffs` method with
`expanded: true` tries to load the content of the diff. After failure it
returns only limited number of changes.

**Solution**

Use `FindChangedPathsRequest` to get the list of changed
files. Add a new method `#changed_files` that returns a list of files
without loading their content.

Changelog: fixed