Add reload button to To-Do list
This allows people to refresh the list manually without the need to
completely refresh the page. It also serves as an unobstrusive loading
indicator in case we do background refreshes where we don't want to hide
the full list of todos.

In order to use this new loading indicator, the refresh of the page in
case you come back to the tab now uses this loading indicator.