Add and backfill work_item_types.old_id column
We are fixing work_item_types table records to have IDs
with fixed values in the range of 1-9.
During the transition, by keeping an old_id column we will
be able to fallback to the old_id when finding these records.

Later we can remove the column and fallbacks in the code.

Changelog: other