Disable the Expect header in QA Maven specs
When publishing a package with $ mvn, an Expect header is used
in all PUT requests. Unfortunately, this doesn't play nice with
some proxies that we have.
We have to disable it. The documented ways to disable this
part on $ mvn commands don't work, so we are overriding
the headers manually to remove the Expect header.