Improve layout of workspace items
- Redesign the User Workspaces UI to simplify how we communicate a
Workspace's state and other important information.
- Create a calculateDisplayFunction function to encapsulate the
Workspace's display state logic. As a result, the implementation of Vue
components that depend on the "display state" is significantly simpler.
The workspace's displayState is calculated based on the
Workspace's actualState and desiredState properties.
It creates an Storybook story that visualizes a Workspace's displayState
for every combination of a Workspace's actualState and desiredState
values.

Changelog: changed
EE: true