Check work_item_epics FF for legacy date rollup
When the hierarchy is not correctly synced to the work item side, the
rolledup date calculation would be not correct.

Since the feature flag `work_items_rolledup_dates` is already by default
enabled, this could lead to dates not being rolled up correctly.

For this reason we add another feature flag check for `work_item_epics`
which is not yet enabled by default.

Changelog: fixed
EE: true