Fetch Arkose Data Exchange payload on Arkose challenge initialization
Payload reuse happens when the Data Exchange payload is part of the data
attributes fetched from the server on page load. The same payload is
reused when the Arkose challenge is re-initialized in the frontend
without page reload (e.g. when user uses "Verify with credit card
instead?" option).

Fetching a new Data Exchange payload when the Arkose challenge is
initialized prevents payload reuse.