Regenerate username if it ends with a file extension
We have a validation in the User model to prevent the creation of
usernames ending with a file extension. However, FFaker sometimes
generates usernames that end with file extensions. This change
modifies a rake task and a fixture to regenerate the username if
it ends with a file extension.

This ensures that all generated usernames comply with our existing
validation rules, maintaining consistency in our user data.