Convert ProjectCacheWorker args to native JSON types
Sidekiq workers use JSON to serialize job arguments. This places some
restrictions on how the classes can be configured and called
https://github.com/sidekiq/sidekiq/wiki/Best-Practices

This includes passing symbol arguments to perform_async and perform_in.
To ensure the arguments are always the same we should stick to passing
strings to perform_now too.

This change converts all of the symbol values for the ProjectCacheWorker