Use nav item title as aria-label on nav link
Currently, the text of any count in a pill on a nav item is part of the
accessible label for a nav item -- for example, if there are 4 issues,
then the accessible label on the "Issues" nav item is "Issues 4." This
is unpredictable and undesired. Instead, the label on the nav item
should match the visual presentation.

Changelog: fixed