Add quickActionsStatus field to create/update note GraphQL mutations
Previously the `errors` field held the results of the quick action,
but this made legitimate messages appear to fail. Now `errors`
is only populated if the quick action failed, and details
are in the `quickActionsStatus` field.

Note that some `errors` will still be blank when calling actions such
as `/unapprove` twice in a row. That's because the quick action
service doesn't not yet have a way to distinguish between some
failures.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/346557

Changelog: fixed