Add rake task to enable work item epics
This adds the following two rake tasks:

- gitlab:work_items:epics:enable
- gitlab:work_items:epics:disable

To either enable or disable the work item epics experience.
When enabling work item epics for the instance, we validate if the data
is in sync. If the data is not in sync, we do not allow to enable the
new experience.

Instances that cannot enable the new experience, need to wait for work
item epics to be GA.

Changelog: added
EE: true