Do not hide To-Do upon state change
This effectively reverts a big chunk of:
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/171318

We keep an individual To-Do around, in order not to force reflows and
trigger unnecessary requests. This allows people to go through the list
one-by-one and mark pending/done in an efficient manner.

As you can just update the To-Do inline, there is no need for individual
"undo" toasts. Let's just get rid of them for now.