Clarify upgrade stop path
This change emphasizes that the path will be predictable in the future.
The intention is to make it bold text so that users that are already used to
reference to this page, now also pay attention to the new guidance. We
could remove the emphasises later.

This also makes it clear that the next planned required stops in the list are
17.8 and 17.11. Before that, it lead readers to believe that any 17 minor.patch
combination would be a required stop, which is wrong.