Stop returning VueWrappers in custom locators
This change removes the capability of our custom locators such as
`findByText` of returning a Vue wrapper and leave the regular DOM
wrapper instead.

Why this change? The method used an internal `__vue__` object to find
the corresponding Vue instance of the element to create the VueWrapper.
This reference is removed in Vue 3 so we can continue to support this.

Tests which made use of these returns have been migrated to use
`findByTestId` which still return the Vue wrapper.