Prevent missing records in Direct Transfer when Sidekiq restarts
In https://gitlab.com/groups/gitlab-org/-/epics/9231, we worked on a few
issues to allow Direct Transfer to handle Sidekiq interruptions and
continue the migration process from where it stopped.

In summary, the implemented solution is to track on Redis the latest
processed line from the NDJSON files, and in case of a job retrial,
Direct Transfer continues from the latest processed line + 1.

However, for some relations, depending on the moment the job is
interrupted, a `PG::UniqueViolation` error may be raised, and usually,
that means some resources were not imported.

This error occurs because we only mark an NDJSON line as processed after
saving all the record and nested records in the database. So, in case of
retry, the same line is processed again, which causes the error.

For example, for each issue in a issues.NDJSON file, Direct Transfer
first saves the issue in the database and then imports the nested
resources (notes, events, milestone, etc). If the job is interrupted
during the import of nested resources when the job is retried,
Direct Transfer will attempt to import the same issue again,
resulting in a `PG::UniqueViolation`, as another issue with the same
IID will already exist and the remaining nested resources won't be
imported.

This changes updates Direct Transfer to delete the previous partially
imported NDJSON object, so the object is fully imported.

Changelog: changed