Use -B gobuildid instead of sha1sum for Workhorse builds
We have some hacks in our Makefile that allow for the generation of
reproducible GNU build IDs. The way this is implemented is kind of
ugly:

1. Build the required Workhorse binaries.

2. To compute a unique and deterministic value for GNU build-id, we
   build the Go binaries a second time.

3. From the first build, we extract its unique and deterministic Go
   build-id, and use that to derive a comparably unique and
   deterministic GNU build-id to inject into the final binary.

4. If we cannot extract a Go build-id, we punt and fallback to using a
   random 32-byte hex string.

This is rather complex and also wasteful, as we essentially have to build
the binaries twice.

To address this, we have upstreamed 089cc68676 (cmd/link: allow deriving
GNU build ID from Go build ID ID, 2023-09-16) that does this for us in
the Go build chain. When passing the `-B gobuildid` linker flag, then
the linker will compute the GNU build ID from the Go build ID and thus
do essentially the same as we did manually in Workhorse. This new feature
has been released with Go v1.22.

Now that we are on Go v1.23, we can finally use this new flag. This
allows us to simplify our build instructions.

In addition, this fixes fixes macOS build failures in Go v1.23.3,
introudced by 54728538 (cmd/link: generate Mach-O UUID when -B flag is
specified). On macOS, builds with our previous scheme failed:

```
link: -B option too long (max 16 digits): 0x<SHA>
```

This change aligns with what we have done in other projects:

- Pages:
  https://gitlab.com/gitlab-org/gitlab-pages/-/merge_requests/1082

- Gitaly: https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7221