Add use_git privilege for Duo Workflows
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/171872 we
introduced the ability to configure Duo Workflow with specific
priviliges for things that the agent is allowed to do. In
https://gitlab.com/gitlab-org/duo-workflow/duo-workflow-service/-/merge_requests/158
we started enforcing those privileges and introduced one extra `use_git`
privilege that was missing on the GitLab side. This adds that privilege
on the GitLab side.