Index migration for SBOM aggregations of severity (NULLS LAST)
Context
--------------

We have an [MR][0] open to start using `NULLS LAST` for the SBOM
dependency page when sorting by `highest_severity` in descending
order.

In a [thread][1] on that MR, it was mentioned that the existing index
would not cover this new sort order, so we need to add an index to
prevent a performance regression

This change
--------------

The `sbom_occurrences` table is a [high traffic table][2], however the
migration style guides only specifically call out using the async
index creation workflow if the migration takes too long:

      If indexing takes a long time to finish (a post-deployment migration
      should take less than 10 minutes) consider indexing asynchronously.[^1]

Therefore, I have added the migration synchronously.

It is also worth noting that the reason we are adding
`highest_severity NULLS FIRST` in the index, even though we are going
to be adding `NULLS LAST` to the actual query, is that the index will
be scanned backwards when the sort order is `DESC`.

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/172050
[1]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/172050#note_2212399479
[2]:https://gitlab.com/gitlab-org/gitlab/-/blob/3d45373454d5a80f8b934902b606b53d77a96eb1/rubocop/rubocop-migrations.yml#L39
[3]:https://docs.gitlab.com/ee/development/database/adding_database_indexes.html#add-an-index-to-improve-existing-queries
[4]:https://docs.gitlab.com/ee/development/migration_style_guide.html#high-traffic-tables

[^1]: I think the guide [here][3] used to explicitly state that
      [high traffic tables][4] by default should have there indexes created
      asynchronously. However, I don't see that called out anymore in the
      docs. I only see the migration timing being called out as a concern.

---

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/172986
Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/502335
Changelog: added