Shorten Vue 3 Jest quarantine check job log
If a merge request touches many files, the `check quarantined` job will
run many specs, most of which are likely to fail. This means the job
output will be large, and likely truncated with the message:

    Job's log exceeded limit of 4194304 bytes.
    Job execution will continue but no more output will be collected.

This means the useful information on how to fix the job is lost.

We now write Jest's output to files, which are exposed as artifacts.
This ensures that the job log will be much smaller, while ensuring that
the job's behaviour can still be debugged.