Pivot "VR in MR" feature from a suggestion to a MR-based approach
Context
-------------

We have an [epic][0] to build a workflow for the
[Duo Vulnerability Resolution][1] feature into the
[Security MR widget][2].

The design for this workflow settled on using in-line suggestions on
an MR to supply the fixes.[^1]

During a limited feature flag roll-out, a [discussion][3] started
around the limitations of a suggestion-based approach for
vulnerabilities that span multiple files.

The result of that discussion was that we should pivot to an MR-based
approach.[^2]

This change
-------------

This is a first iteration to pivot to a workflow of creating a
separate MR rather than using suggestions.

It re-uses most of the suggestion-based workflow's code, modifying it
to:

- pipe the LLM response into our MR creation service as opposed to a
  suggestion creation service
- has this new MR use the feature branch as its target branch, rather
   than the default branch
- remove "suggestion" related nomenclature
- adds some spec coverage around the new flow

This feature is still behind a [feature flag][4] and we we continue
iterating based on [user acceptance testing][5]

[0]:https://gitlab.com/groups/gitlab-org/-/epics/14862
[1]:https://docs.gitlab.com/ee/user/application_security/vulnerabilities/index.html#vulnerability-resolution
[2]:https://docs.gitlab.com/ee/user/application_security/#merge-request
[3]:https://gitlab.com/gitlab-org/gitlab/-/issues/502919
[4]:https://gitlab.com/gitlab-org/gitlab/-/issues/482753
[5]:https://gitlab.com/gitlab-org/gitlab/-/issues/500101
[6]:https://gitlab.com/gitlab-org/gitlab/-/issues/502919#note_2201105477

[^1]:https://gitlab.com/gitlab-org/gitlab/-/issues/407275/designs/design_1722547971389.png
[^2]: That decision is discussed [here][6] and also includes the
      requirement of keeping the original release timeline

---

epic: https://gitlab.com/groups/gitlab-org/-/epics/14862
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/172185
resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/503403
Changelog: changed
EE: true

Co-authored-by: Darby Frey <dfrey@gitlab.com>
Co-authored-by: Michael Becker <mbecker@gitlab.com>