Replace jackspeak dependency with @gitlab/noop
The jackspeak dependency is a dependency of the glob package. Inside the
glob package it is only used in the CLI features of the package, which
we make no use of. Some decisions of the authors (not questioning them)
have led to weird behaviors of yarn@v1:
https://github.com/isaacs/jackspeak/issues/5

With this we are replacing the the problematic dependency subtree with
`@gitlab/noop` which is simply an empty package with no dependencies. Of
all the other package managers, only `pnpm` has a native
dependency-removal and that also rather recently, so this empty package
is a smug workaround.

See also:
- https://pnpm.io/package_json#pnpmoverrides
- https://gitlab.com/gitlab-org/frontend/noop