Fix epic imports missing Epic Work Item info
When importing epic work items, information that was added
to epic work items is missing while older information added
to the legacy epic is being imported.

The problem is related to the fact that even when importing epic
work item we in fact are importing legacy Epic and consequently
we create the related Epic Work Item record. However when we import
the legacy Epic record we preload associations on the legacy
Epic model, which results in not loading information from
the corresponding Epic Work Item.

This change removes the association preloading when importing epics
to fix this problem.

We will be moving the export/import of epics to work with Epic
Work Item once Epic Work Items become GA and we can fix this code back
at that time.