Invalidate branch api cache when protected branches updated
Resolves https://gitlab.com/gitlab-org/gitlab/-/issues/485513+

This adds the project's cache key to the branches cache context.

The projects cache key includes the updated at timestamp.

We touch the project when we modify protected branches at the project
level. This change also makes sure we touch all projects in a group when
refreshing the protected branches cache.