Update ReferenceRewriter to handle projects and groups
Updated specs into a table syntax for easier consumption.
Table syntax tests only valid references so old spec examples
for invalid references are still left in place for spec coverage.

Changes in this MR are intended to allow ReferenceRewriter to rewrite
references in the markdown content while the content is being moved
across different namespaces.


Previously we had "move/clone" functionality for issues only.
Issues used to be a project level only item. Now as we transition
to WorkItem we also have group level work items, e.g. Epic Work Item,
and in order to keep parity on "move" functionality we need to be
able to also move group level WorkItems(Issues). So this MR adapts
the ReferenceRewriter to handle references in markdown when items
are moved across groups and projects.

As part of this improvement I had to also change the
IssueReferenceFilter so that it does not passes on group level
items, in order for all group level items to be handled as Work Item,
but WorkItemReferenceFilter. Otherwise if references are detected
by IssueReferenceFilter those are not being picked up by
WorkItemReferenceFilter.