Log requested scopes when auth is denied
In production, we are seeing some log messages show that
authentication in the `meta.auth_fail_reason` field was denied for the
requested scopes even though the API request succeeded.

It's possible that auth was requested for one set of scopes,
but succeeded for another. This commit adds
`meta.auth_fail_requested_scopes` to better understand which
scopes were not available.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/505746

Changelog: changed