Fix code reloading problems with custom database types
As described in https://github.com/rails/rails/issues/27337,
registering a custom, autoloaded `ActiveRecord::Type` in an
initializer is not safe for code reloading. Previously when
`Gitlab::Database::Type::IndifferentJsonb` was registered in the
ActiveRecord the registry and code reloading occurred, a `TypeError`
would occur because the registry had an old reference.

Let's avoid this code reloading issue by not using the ActiveRecord
registry and instantiate the types explicitly.

Relates to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/2051

Changelog: fixed