Revert "fixing markdown formatting for pytest-cov coverage command example"
This reverts commit a67719e25afba516d43dd53ff647cd2fde19c4e0.

It seems there is some confusion between this change and
232538cb3c9f130e124c72da45136ff1008fe56d.  I think the issue is that
the \| renders differently in the generated documentation site vs.
viewing the .md file via the gitlab file browser.

Currently, the regex renders with a literal \| when veiwed on
https://docs.gitlab.com/ee/ci/testing/code_coverage.html, making this
not actually match the pycov output.  It looks correct on
https://gitlab.com/gitlab-org/gitlab/-/blob/master/doc/ci/testing/code_coverage.md?ref_type=heads
-- but I don't think most people are looking at the raw pages in the
repo.

If we have to choose, I think the docs site should render correctly.
To avoid confusion, I've put a comment in.