Fix 500 error for /api/v4/groups?all_available=false&statistics=true API
The GET /api/v4/groups?all_available=false&statistics=true API was
throwing internal server error when logged in using an admin user
because the generated query was throwing the following error:

```
PG::UndefinedTable: ERROR:  invalid reference to FROM-clause entry for table "namespaces"
LINE 1: ...tories" ON "group_wiki_repositories"."group_id" = "namespace...
                                                             ^
DETAIL:  There is an entry for table "namespaces", but it cannot be referenced from this part of the query.
```

Here we generate the semantically correct query and add the request
specs to avoid a regression again.

Changelog: fixed