Fail when removed specs are not removed from quarantine
There' a gap in the logic of the `jest vue3 check quarantined` CI job:
it doesn't fail when a spec that's listed in the quarantine file is
deleted.

This is because the job works by running specs listed in the quarantine
file and failing if any spec passes. But, if a spec no longer exists, it
cannot be run, so it cannot pass, and therefore cannot fail the job.

Now we ensure that removed spec files that are listed in the quarantine
list also fail the job.