Fix dependency on ApplicationSetting stubs for `let_it_be`
Usages of let_it_be fail due to application setting not been stubbed
yet. The particular case was related to the use of `trace_live` trait
of the `Ci::Build` factory, which tries to hide secrets from a trace,
causing a lookup into ApplicationSetting#ci_job_token_signing_key which
at the time is `nil` due to not being stubbed yet. This causes the
CI_JOB_TOKEN to be `nil` by consequence.

Changelog: fixed