Remove unsused GitLab Shell mixin
This mixin provide a shortcut to Gitlab::Shell instance
by gitlab_shell. This Gitlab::Shell class is an artifact
of a time when common repository operations were performed
by calling out to scripts in the gitlab-shell project. Now,
these operations are all performed by Gitaly, and are mostly
accessible through the Repository class.