Fix 403 errors when setting CI/CD variables via REST API
This MR updates the API authorization logic for the CI/CD
settings of projects and groups. Previously, we were only
checking a user was able to admin the group or project and
authorization would fail if they were assigned a custom role
with the admin_cicd_variables permission.

Changelog: fixed