Temporary fix for activity feed tooltip contrast
We can't use `text-subtle` in a tooltip currently in light mode, becuase
`text-subtle` assumes it will appear on one of our light mode
backgrounds. This means that currently the subtle text on the second
line of an activity feed tooltip is unreadable in light mode.

This MR temporarily adjusts this so that the second line has a reduced
opacity instead, so that it can be made similarly styled but readable in
both light and dark modes.

Changelog: fixed