Skip GraphQL CSRF check when not session authed
We only need to verify the CSRF token when the user is authenticated via
a session cookie. This also makes it similar to other controllers where
we raise an exception when the token does not match.

This fixes a race condition when restoring multiple tabs using the
remember_me token where a CSRF mismatch can result to the user being
logged out.