Improves Service Desk list query and includes tickets
In projects with lots of issues the query for the
Service Desk list view is slow because we filter
by the author username of the Support bot.

Now we directly filter by the user id and
work item type id.

Additionally we include work items of type ticket
so the Service Desk list app can also consume tickets.
This is necessary because the frontend app fetches
issues via GraphQL instead of work items. Once we
migrated Service Desk issues to work items of type
ticket we can safely refactor the list app to fetch
work items via GraphQL.