Skip validations when transforming references
We see some errors attempting to save objects which fail a
validation that would be switched off when `importing?`. For example:

https://gitlab.com/gitlab-org/gitlab/-/blob/2f369a0339caddc55a736bfc74abb0064ac48550/app/models/diff_note.rb#L24

The objects we are saving were initially created with the `importing?`
set to `true`, so here we continue to set it to allow the objects to
skip those same validations when their references are being updated.

Changelog: fixed